backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
playwright==1.40.0
crewai==0.28.0
langchain>=0.1.10,<0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
httpx==0.25.0
aiofiles==23.2.1
websockets==12.0
